+++ databento/common/enums.py
@@ -6,17 +6,6 @@ from enum import IntFlag
 from enum import unique
 from typing import Callable, TypeVar
 
-from databento_dbn import ImbalanceMsg
-from databento_dbn import InstrumentDefMsg
-from databento_dbn import MBOMsg
-from databento_dbn import MBP1Msg
-from databento_dbn import MBP10Msg
-from databento_dbn import OHLCVMsg
-from databento_dbn import StatMsg
-from databento_dbn import TradeMsg
-
-from databento.live import DBNRecord
-
 
 M = TypeVar("M", bound=Enum)
 
@@ -71,7 +60,7 @@ def coercible(enum_type: type[M]) -> type[M]:
 
     def coerced_new(enum: type[M], value: object) -> M:
         if value is None:
-            raise TypeError(
+            raise ValueError(
                 f"value `{value}` is not coercible to {enum_type.__name__}.",
             )
         try:
@@ -111,7 +100,6 @@ class StringyMixin:
             return getattr(self, "name").lower()
         return getattr(self, "value")
 
-
 @unique
 @coercible
 class HistoricalGateway(StringyMixin, str, Enum):
@@ -143,77 +131,6 @@ class Dataset(StringyMixin, str, Enum):
 
     GLBX_MDP3 = "GLBX.MDP3"
     XNAS_ITCH = "XNAS.ITCH"
-
-
-@unique
-@coercible
-class Schema(StringyMixin, str, Enum):
-    """
-    Represents a data record schema.
-    """
-
-    MBO = "mbo"
-    MBP_1 = "mbp-1"
-    MBP_10 = "mbp-10"
-    TBBO = "tbbo"
-    TRADES = "trades"
-    OHLCV_1S = "ohlcv-1s"
-    OHLCV_1M = "ohlcv-1m"
-    OHLCV_1H = "ohlcv-1h"
-    OHLCV_1D = "ohlcv-1d"
-    DEFINITION = "definition"
-    IMBALANCE = "imbalance"
-    STATISTICS = "statistics"
-
-    def get_record_type(self) -> type[DBNRecord]:
-        if self == Schema.MBO:
-            return MBOMsg
-        if self == Schema.MBP_1:
-            return MBP1Msg
-        if self == Schema.MBP_10:
-            return MBP10Msg
-        if self == Schema.TBBO:
-            return MBP1Msg
-        if self == Schema.TRADES:
-            return TradeMsg
-        if self == Schema.OHLCV_1S:
-            return OHLCVMsg
-        if self == Schema.OHLCV_1M:
-            return OHLCVMsg
-        if self == Schema.OHLCV_1H:
-            return OHLCVMsg
-        if self == Schema.OHLCV_1D:
-            return OHLCVMsg
-        if self == Schema.DEFINITION:
-            return InstrumentDefMsg
-        if self == Schema.IMBALANCE:
-            return ImbalanceMsg
-        if self == Schema.STATISTICS:
-            return StatMsg
-        raise NotImplementedError(f"No message type for {self}")
-
-
-@unique
-@coercible
-class Encoding(StringyMixin, str, Enum):
-    """
-    Represents a data output encoding.
-    """
-
-    DBN = "dbn"
-    CSV = "csv"
-    JSON = "json"
-
-
-@unique
-@coercible
-class Compression(StringyMixin, str, Enum):
-    """
-    Represents a data compression format (if any).
-    """
-
-    NONE = "none"
-    ZSTD = "zstd"
 
 
 @unique
@@ -251,19 +168,6 @@ class Delivery(StringyMixin, str, Enum):
     DOWNLOAD = "download"
     S3 = "s3"
     DISK = "disk"
-
-
-@unique
-@coercible
-class SType(StringyMixin, str, Enum):
-    """
-    Represents a symbology type.
-    """
-
-    INSTRUMENT_ID = "instrument_id"
-    RAW_SYMBOL = "raw_symbol"
-    PARENT = "parent"
-    CONTINUOUS = "continuous"
 
 
 @unique