+++ databento/common/parsing.py
@@ -7,8 +7,8 @@ from functools import singledispatch
 from numbers import Number
 
 import pandas as pd
+from databento_dbn import SType
 
-from databento.common.enums import SType
 from databento.common.symbology import ALL_SYMBOLS
 from databento.common.validation import validate_smart_symbol