+++ databento/historical/api/timeseries.py
@@ -4,13 +4,13 @@ from datetime import date
 from os import PathLike
 
 import pandas as pd
+from databento_dbn import Compression
+from databento_dbn import Encoding
+from databento_dbn import Schema
+from databento_dbn import SType
 
 from databento.common.dbnstore import DBNStore
-from databento.common.enums import Compression
 from databento.common.enums import Dataset
-from databento.common.enums import Encoding
-from databento.common.enums import Schema
-from databento.common.enums import SType
 from databento.common.parsing import datetime_to_string
 from databento.common.parsing import optional_datetime_to_string
 from databento.common.parsing import optional_symbols_list_to_string