+++ databento/live/gateway.py
@@ -4,12 +4,14 @@ import dataclasses
 import logging
 from functools import partial
 from io import BytesIO
+from operator import attrgetter
 from typing import TypeVar
 
+from databento_dbn import Encoding
+from databento_dbn import Schema
+from databento_dbn import SType
+
 from databento.common.enums import Dataset
-from databento.common.enums import Encoding
-from databento.common.enums import Schema
-from databento.common.enums import SType
 
 
 logger = logging.getLogger(__name__)
@@ -53,11 +55,9 @@ class GatewayControl:
             ) from type_err
 
     def __str__(self) -> str:
-        tokens = "|".join(
-            f"{k}={str(v)}"
-            for k, v in dataclasses.asdict(self).items()
-            if v is not None
-        )
+        fields = tuple(map(attrgetter("name"), dataclasses.fields(self)))
+        values = tuple(getattr(self, f) for f in fields)
+        tokens = "|".join(f"{k}={v}" for k, v in zip(fields, values) if v is not None)
         return f"{tokens}\n"
 
     def __bytes__(self) -> bytes: